v0.0.8 holiday updates (#500)
* formatting JS

* audio config script

very rough draft of a script to configure .asoundrc very easily

* Update conf-audio.sh

removed some of the messy echos in the code

* Update conf-audio.sh

edited the file a little too quick before running out to dinner... its fixed now...

* added tvshows service

fixed spacing

* removed  param, added configurable refresh interval

* add capability to use multiple hotwords for activation. fixes #442

* formatting config

* fix rss config read error. resolves #409

* fix traffic errors. don't make requests with empty key string

* 0.0.6

* add basic client side error logging

* protip for validating config

* fix reminders. resolves #438

* Update README.md

added Feature Requests and Suggestions to the readme.md...

* Update README.md

* Update README.md

* Bug Fix #457 (#458)

HOTFIX #457

* set a default format language as an english because bing maps API wnats the value of 'dateTime' as 'PM' or 'AM'

* use sonus ^0.1.2 fixes #462

* Add pt-BR Locale

* use https on YouTube #471

* add Johnny five support (#464)

* adding johnny-five functionality for motion detection

* quick fix missing comma

* add info to config.example.js

* quick fix for change config.motion.type to config.motion.mode

* add commas

* add ipcRenderer to autosleep.js

* add spawn commands for motion spotter

* ipcrenderer fix

* further ipcrendering attempts

* ipcrenderer fix

* add script for monitor on/off.

* add script

* Update pi-install.sh

* remove some of the console.errors prior to ipc send... add calibrated IPC

* Add Enabled/disabled function...

removed model and mode from config since proximity will not function properly.

* Add fix for voice commands
in en.json regarding autosleep

* Update package.json

updated `package.json` to match the current file in `dev`

* Update package.json

* add comma

* Update config.example.js

* Update package.json

* Update config.example.js

* Smart mirror patch 0.0.7

* manual changes from #479 (#483)

* manual changes from #479

spacing

* agnostic autoTimer at exit

auto timer

* remove projectId requirement in config. (#484)

* remove projectId requirement in config.

sonus.js will now use the keyfile to load the project ID.

sonus.js will now check to make sure the model file exists and default to snowboy.umdl if it does not.

* fix config.language

* Update sonus.js

* corrected config.js path on line 4

* hotword helper function

* initial refactor

breaking out view components and refactoring weather

about halfway through refactoring

about halfway through refactoring

fixing timer

fixing timer

rss

rss

soundcloud

scrobble

clean up and finish refactor

removing dupe

mend

* use plugin tree for dir structure

* moving pt local

fixing qr plugin

* Smart mirror remote (#450)

* add voice remote for mirror

add connected/disconnected events

* remote devtools kiosk reload

* expose mirror remote IP and port

* expose mirror remote IP and port

* remote connection state management

* add visual breakpoint for JS in index.html

* Adding QR functionality using googleapis charts

this will need to be revised as google can end use of this as it is depreciated...  This is more to provide a way of testing the concept and not delay development of the config pages...

* Committing remaining changes for QRservice

* Patch conflicts

* Update config.js

* Delete config.example.js

* Add files via upload

* Update index.html

* missed a <<<

* moving pt local

fixing qr plugin

* merge configUI from bill

renaming

disable motion by default

move master to default config name

fixing default

patch index

removing config.js

re-adding fs to index.html

config fix

removing js

finishing sonus merge

merging configUI from bill

* Patch Sleep Functionality

 *  removed "Screen On" & "Screen Off" voice commands
 *  removed "Wake Up" voice command
 *  "Sleep" voice command now calls AutoSleepService.sleep function
* Hotword now wakes the mirror by turning on screen if it is sleeping.

* config schema per plugin

* add reboot functionality for remote

* adding maker channel (HomeAutomation, IFTTT, Custom HTTP Requests, etc)

* Update config.schema.json

Prepared config.schema.json for next push from `justbill2020/config-form-improvements`

* squash commit from justbill2020/config-form-improvements

Major changes outlined below:

1) added order key to form[] array so that general, speech, and remote
   are at the top of the config form. also allows for more defined
   ordering of form elements in the future.

2) added config.schema.json for tvshows plugin.

3) manually defined "remote{}" to config in main.js if config.json is
   not found. this can be done other ways... if config.json is not
   found currently it will display a white screen. this should instead
   display the qrcode and remote link. stopped working on forcing that,
   will leave getting that to work to Evan.

4) fixed variable used to merge form[] and value{} from each config.schema.js

5) configFN changed to configPath

6) don't app.quit on failing to load config.json

* adding smart mirror first run

* when a new config is written, relaunch

* ensure config works after configuration

* mobile style fix

* fix autosleep variables wakeCmd and sleepCmd

* motion fix & looking into RSS issue

* get more error info on config.html ...

...when jsonForm fails

* move deps for jsonForm to https://gitcdn.xyz

* use sonus 0.1.4+

* 0.0.8

* Readme update for discourd and docs

remote colin in readme

spelling

* remote menu and ui fixes

* Update config.schema.json

* simple script to agrigate config files

better file names

* use videoSyndicated for YouTube search

* remove npm-platform-dependencies

* fixing merge issue

* some cleanup

* Updating discord link in README.

* remove random jQuery include

* Update remote.js

* Center suggestions

+++ main.js
@@ -3,6 +3,8 @@
 const electron = require('electron')
 // Child Process for keyword spotter
 const {spawn, exec} = require('child_process')
+// Smart mirror remote
+const remote = require('./remote.js')
 // Module to control application life.
 const app = electron.app
 // Module to create native browser window.
@@ -12,26 +14,23 @@ const powerSaveBlocker = electron.powerSaveBlocker
 powerSaveBlocker.start('prevent-display-sleep')
 
 // Launching the mirror in dev mode
-const DevelopmentMode = process.argv[2] == "dev";
+const DevelopmentMode = process.argv[2] == "dev"
 
 // Load the smart mirror config
-var config;
+let config
 try {
-  config = require(__dirname + "/config.js");
+  config = require("./config.json")
 } catch (e) {
-  var error = "Unknown Error"
-
+  let error = "Unknown Error"
+  config = require("./config.default.json")
   if (typeof e.code != 'undefined' && e.code == 'MODULE_NOT_FOUND') {
     error = "'config.js' not found. \nPlease ensure that you have created 'config.js' " +
       "in the root of your smart-mirror directory."
   } else if (typeof e.message != 'undefined') {
     console.log(e)
     error = "Syntax Error. \nLooks like there's an error in your config file: " + e.message + '\n' +
-    'Protip: You might want to paste your config file into a JavaScript validator like http://jshint.com/'
+      'Protip: You might want to paste your config file into a JavaScript validator like http://jshint.com/'
   }
-
-  console.log("Config Error: ", error)
-  app.quit()
 }
 
 // Keep a global reference of the window object, if you don't, the window will
@@ -41,17 +40,17 @@ let mainWindow
 function createWindow() {
 
   // Get the displays and render the mirror on a secondary screen if it exists
-  var atomScreen = electron.screen;
-  var displays = atomScreen.getAllDisplays();
-  var externalDisplay = null;
+  var atomScreen = electron.screen
+  var displays = atomScreen.getAllDisplays()
+  var externalDisplay = null
   for (var i in displays) {
     if (displays[i].bounds.x > 0 || displays[i].bounds.y > 0) {
-      externalDisplay = displays[i];
-      break;
+      externalDisplay = displays[i]
+      break
     }
   }
 
-  var browserWindowOptions = { width: 800, height: 600, icon: 'favicon.ico', kiosk: true, autoHideMenuBar: true, darkTheme: true };
+  var browserWindowOptions = { width: 800, height: 600, icon: 'favicon.ico', kiosk: true, autoHideMenuBar: true, darkTheme: true }
   if (externalDisplay) {
     browserWindowOptions.x = externalDisplay.bounds.x + 50
     browserWindowOptions.y = externalDisplay.bounds.y + 50
@@ -65,7 +64,7 @@ function createWindow() {
 
   // Open the DevTools if run with "npm start dev"
   if (DevelopmentMode) {
-    mainWindow.webContents.openDevTools();
+    mainWindow.webContents.openDevTools()
   }
 
   // Emitted when the window is closed.
@@ -78,25 +77,111 @@ function createWindow() {
 }
 
 // Initilize the keyword spotter
-var kwsProcess = spawn('node', ['./sonus.js'], {detached: false})
-// Handel messages from node
-kwsProcess.stderr.on('data', function (data) {
-  var message = data.toString()
-  console.error("ERROR", message.substring(4))
-})
+if (config && config.speech) {
+  var kwsProcess = spawn('node', ['./sonus.js'], { detached: false })
+  // Handel messages from node
+  kwsProcess.stderr.on('data', function (data) {
+    var message = data.toString()
+    console.error("ERROR", message.substring(4))
+  })
 
-kwsProcess.stdout.on('data', function (data) {
-  var message = data.toString()
-  if (message.startsWith('!h:')) {
-    mainWindow.webContents.send('hotword', true)
-  } else if (message.startsWith('!p:')) {
-    mainWindow.webContents.send('partial-results', message.substring(4))
-  } else if (message.startsWith('!f:')) {
-    mainWindow.webContents.send('final-results', message.substring(4))
-  } else {
-    console.error(message.substring(3))
+  kwsProcess.stdout.on('data', function (data) {
+    var message = data.toString()
+    if (message.startsWith('!h:')) {
+      mainWindow.webContents.send('hotword', true)
+    } else if (message.startsWith('!p:')) {
+      mainWindow.webContents.send('partial-results', message.substring(4))
+    } else if (message.startsWith('!f:')) {
+      mainWindow.webContents.send('final-results', message.substring(4))
+    } else {
+      console.error(message.substring(3))
+    }
+  })
+}
+// Motion detection
+if (config && config.motion && config.motion.enabled) {
+  var mtnProcess = spawn('npm', ['run', 'motion'], { detached: false })
+  // Handel messages from node
+  mtnProcess.stderr.on('data', function (data) {
+    var message = data.toString()
+    console.error("ERROR", message.substring(4))
+  })
+
+  mtnProcess.stdout.on('data', function (data) {
+    var message = data.toString()
+    if (message.startsWith('!s:')) {
+      console.log(message.substring(3))
+      mainWindow.webContents.send('motionstart', true)
+    } else if (message.startsWith('!e:')) {
+      console.log(message.substring(3))
+      mainWindow.webContents.send('motionend', true)
+    } else if (message.startsWith('!c:')) {
+      console.log(message.substring(3))
+      mainWindow.webContents.send('calibrated', true)
+    } else if (message.startsWith('!E:')) {
+      console.log(message.substring(3))
+      mainWindow.webContents.send('Error', message.substring(3))
+      mtnProcess.kill();
+    } else {
+      console.error(message)
+    }
+  })
+}
+
+if (config.remote && config.remote.enabled) {
+  remote.start()
+
+  // Deturmine the local IP address
+  const interfaces = require('os').networkInterfaces()
+  let addresses = []
+  for (let k in interfaces) {
+    for (let k2 in interfaces[k]) {
+      let address = interfaces[k][k2]
+      if (address.family === 'IPv4' && !address.internal) {
+        addresses.push(address.address)
+      }
+    }
   }
-})
+  console.log('Remote listening on http://%s:%d', addresses[0], config.remote.port)
+
+  remote.on('command', function (command) {
+    mainWindow.webContents.send('final-results', command)
+  })
+
+  remote.on('connected', function () {
+    mainWindow.webContents.send('connected')
+  })
+
+  remote.on('disconnected', function () {
+    mainWindow.webContents.send('disconnected')
+  })
+
+  remote.on('devtools', function (open) {
+    if (open) {
+      mainWindow.webContents.openDevTools()
+    } else {
+      mainWindow.webContents.closeDevTools()
+    }
+  })
+
+  remote.on('kiosk', function () {
+    if (mainWindow.isKiosk()) {
+      mainWindow.setKiosk(false)
+    } else {
+      mainWindow.setKiosk(true)
+    }
+  })
+
+  remote.on('reload', function () {
+    mainWindow.reload()
+  })
+
+  remote.on('relaunch', function() {
+    console.log("Relaunching...")
+    app.relaunch()
+    app.quit()
+  })
+}
 
 // Motion detection
 if(config.motion && config.motion.enabled){
@@ -140,14 +225,15 @@ app.on('window-all-closed', function () {
 
 // No matter how the app is quit, we should clean up after ourselvs
 app.on('will-quit', function () {
-  kwsProcess.kill()
-
+  if (kwsProcess) {
+    kwsProcess.kill()
+  }
   // While cleaning up we should turn the screen back on in the event 
   // the program exits before the screen is woken up
-  if(mtnProcess){
+  if (mtnProcess) {
     mtnProcess.kill()
   }
-  if(config.autoTimer && config.autoTimer.wake_cmd){
-    exec(config.autoTimer.wake_cmd).kill()
+  if (config.autoTimer && config.autoTimer.wakeCmd) {
+    exec(config.autoTimer.wakeCmd).kill()
   }
 })
